Refactor: Move momentShim to a new location and update import paths
