+++ typescript/api/services/RDMPService.ts
@@ -30,7 +30,7 @@ import {
   Sails,
   Model
 } from "sails";
-import moment from '../helpers/momentShim';
+import moment from '../shims/momentShim';
 import numeral from 'numeral';
 
 import {
