+++ typescript/api/services/RaidService.ts
@@ -27,7 +27,7 @@ import {
 } from "sails";
 import { RaidApi, RaidCreateRequest, Title, ModelDate, Description, Access, AlternateUrl, Contributor, ContributorRoleCreditNisoOrgType, ContributorRoleSchemeType, Organisation } from '@researchdatabox/raido-openapi-generated-node';
 
-import moment from '../helpers/momentShim';
+import moment from '../shims/momentShim';
 import numeral from 'numeral';
 import axios from 'axios';
